temp save

+++ cogs/Voice.py
@@ -1,19 +1,75 @@
 from discord import ApplicationContext, option
 from discord.ext.commands import Bot
 
-from Assets import assets
+import Assets
 from YTDL import *
 
-res = assets()
+res = Assets.assets()
+
+
+class VoiceClients:
+    def __init__(self, voice, volume):
+        self.source_queue = []
+        self.player_queue = []
+        self.voice = voice
+        self.volume = volume
+        self.loop = False
+        self.is_playing = False
+
+    def add_song(self, player, source):
+        self.player_queue.append(player)
+        self.source_queue.append(source)
+        return self
+
+    def set_volume(self, volume):
+        self.volume = volume
+        self.voice.source.volume = volume
+        return self
+
+    def toggle_loop(self):
+        self.loop = not self.loop
+        return self
+
+    def move_to(self, voice):
+        self.voice = voice
+        return self
+
+    def play(self):
+        print('play')
+        self.is_playing = True
+        return self
+
+    def pause(self, error=None):
+        print('pause', error if error is not None else '')
+        self.is_playing = False
+        return self
+
+    def purge(self):
+        self.player_queue = []
+        self.source_queue = []
+        return self
+
+    def head(self):
+        return self.player_queue[0], self.source_queue[0]
+
+    def pop2(self):
+        self.player_queue.pop(0)
+        self.source_queue.pop(0)
+        return self.player_queue[0], self.source_queue[0]
+
+    def has_next(self):
+        return True if len(self.player_queue) > 1 else False
+
+    def is_empty(self):
+        return True if len(self.player_queue) < 0 else False
 
 
 class Voice(commands.Cog):
 
     def __init__(self, bot: Bot):
         self.bot = bot
-        self.source_queue = []
-        self.player_queue = []
-        self.volume = 1
+        self.voice_clients = dict()
+        self.default_volume = 100
 
     @commands.slash_command(ignore_extra=False, help='Joins the channel you are in', aliases=['Join'])
     @option('channel',
@@ -27,48 +83,49 @@ class Voice(commands.Cog):
             input_type=bool,
             required=False)
     async def join(self, ctx: ApplicationContext, channel, message):
+        await ctx.response.defer()
         if channel is None:
             try:
                 channel = ctx.author.voice.channel
-                print(channel)
             except:
                 await ctx.respond('Please join a channel or give a channel to join')
                 return None
 
         else:
-            print('f')
-            for ch in ctx.guild.voice_channels:
-                if channel == ch.name:
-                    channel = ch
-                    break
-                else:
-                    await ctx.respond('Please enter a valid voice channel')
-                    return
+            if channel not in [x.name for x in ctx.guild.voice_channels]:
+                await ctx.followup.send('Please enter a valid voice channel')
+                return None
+            else:
+                for ch in ctx.guild.voice_channels:
+                    if channel == ch.name:
+                        channel = ch
+                        break
 
         if ctx.voice_client is not None:
-            print('mov')
             await ctx.voice_client.move_to(channel)
+            self.voice_clients[ctx.guild.id].move_to(ctx.voice_client)
         else:
-            print('conn', channel.guild)
             await channel.connect()
+            self.voice_clients[ctx.guild.id] = VoiceClients(ctx.voice_client, self.default_volume)
 
         if message:
-            await ctx.respond('Joined voice channel')
+            await ctx.followup.send('Joined voice channel')
             return None
         return ctx
 
     @commands.slash_command(help='Leaves all channels')
     async def leave(self, ctx: ApplicationContext):
-        await ctx.voice_client.disconnect(force=True)
+        if ctx.voice_client is not None:
+            await ctx.voice_client.disconnect(force=True)
+            self.voice_clients.pop(ctx.guild.id)
         await ctx.respond('Left voice channel')
-        await self.purge(None)
 
     @commands.slash_command()
     async def volume(self, ctx, volume):
         volume = int(volume)
-        self.volume = volume / 100
+        # TODO change guild volume in json parameters
         if ctx.voice_client is not None:
-            ctx.voice_client.source.volume = volume / 100
+            self.voice_clients[ctx.guild.id].set_volume(volume)
         await ctx.respond(f'Changed volume to {volume}')
 
     @commands.slash_command(help='farts in the current channel or join the channel you are in')
@@ -77,9 +134,14 @@ class Voice(commands.Cog):
             ctx = await self.join(ctx, None, False)
         if ctx is None:
             return
-        source = discord.PCMVolumeTransformer(discord.FFmpegPCMAudio(res + f'fart/fart{random.randrange(9)}.mp3'))
-        ctx.voice_client.play(source, after=lambda e: print(f'Player error: {e}') if e else None)
-        await ctx.delete(delay=0)
+
+        voice_client = self.voice_clients[ctx.guild.id]
+        if not voice_client.is_playing:
+            source = discord.PCMVolumeTransformer(discord.FFmpegPCMAudio(res + f'fart/fart{random.randrange(9)}.mp3'))
+            voice_client.play()
+            voice_client.voice.play(source, after=lambda e: (print(f'Player error: {e}'), voice_client.pause()) if e
+                                    else voice_client.pause())
+            await ctx.delete(delay=0)
 
     @commands.slash_command()
     async def info(self, ctx, url, message=True):
@@ -97,8 +159,32 @@ class Voice(commands.Cog):
 
     @commands.slash_command()
     async def play(self, ctx, url):
+        def play_next(error):
+            print(error)
+            voice_client = self.voice_clients[ctx.guild.id]
+            voice_client.pause()
+            if voice_client.is_empty():
+                print('empty')
+                return
+            if voice_client.loop:
+                print('loop')
+                player, source = voice_client.head()
+            elif voice_client.has_next():
+                print('next')
+                player, source = voice_client.pop2()
+            else:
+                print('else')
+                return
+
+            voice_client.play()
+            voice_client.voice.play(player, after=play_next)
+            send = ctx.send(f'Now playing: ***{player.title}***\nFrom {source}')
+            asyncio.run_coroutine_threadsafe(send, self.bot.loop)
+
         if not self.bot.voice_clients:
             ctx = await self.join(ctx, None, message=False)
+        else:
+            await ctx.response.defer()
         if ctx is None:
             return
 
@@ -117,81 +203,66 @@ class Voice(commands.Cog):
                     player = infos[0]
                     source = 'Youtube'
 
-            self.player_queue.append(player)
-            self.source_queue.append(source)
-            if ctx.voice_client.is_playing():
-                await ctx.respond(f'Added ***{player.title}*** to the queue')
+            voice_client = self.voice_clients[ctx.guild.id]
+            voice_client.add_song(player, source)
+            if voice_client.is_playing:
+                await ctx.followup.send(f'Added ***{player.title}*** to the queue')
                 return
             else:
-                await ctx.voice_client.play(player, after=lambda e: self.play_next(ctx))
-                await ctx.respond(f'Now playing: ***{player.title}***\nFrom {source}')
+                voice_client.play()
+                voice_client.voice.play(player, after=play_next)
+                await ctx.followup.send(f'Now playing: ***{player.title}***\nFrom {source}')
                 return
 
     @commands.slash_command(help='stops the music playing')
     async def stop(self, ctx):
         ctx.voice_client.stop()
-        self.player_queue = []
-        self.source_queue = []
+        self.voice_clients[ctx.guild.id].pause().purge()
         await ctx.respond('Stopped voice and purged queue')
 
     @commands.slash_command()
     async def pause(self, ctx):
-        ctx.voice_client.pause()
+        self.voice_clients[ctx.guild.id].pause().\
+            voice().pause()
         await ctx.respond('Paused voice')
 
     @commands.slash_command()
     async def resume(self, ctx):
-        ctx.voice_client.resume()
+        self.voice_clients[ctx.guild.id].play().\
+            voice().resume()
         await ctx.respond('Resuming')
 
     @commands.slash_command()
     async def loop(self, ctx):
         if ctx.voice_client is not None:
-            ctx.voice_client.loop = not ctx.voice_client.loop
-            await ctx.respond('Loop is on' if ctx.voice_client.loop else 'Loop is off')
+            voice_client = self.voice_clients[ctx.guild.id]
+            voice_client.toggle_loop()
+            await ctx.respond('Loop is on' if voice_client.loop else 'Loop is off')
         else:
+            # TODO change guild settings in json
             await ctx.respond('Bot must be connected to a voice channel to enable looping')
-
-    def play_next(self, ctx):
-        if len(self.player_queue) > 0:
-            if not ctx.voice_client.loop:
-                self.player_queue.pop(0)
-                self.source_queue.pop(0)
-
-        if len(self.player_queue) > 0:
-            player = self.player_queue[0]
-            source = self.source_queue[0]
-            ctx.voice_client.stop()
-            asyncio.run_coroutine_threadsafe(ctx.respond(f'Now playing: ***{player.title}***\nFrom {source}'),
-                                             self.bot.loop)
-            ctx.voice_client.play(player, after=lambda e: (self.play_next(ctx)))
 
     @commands.slash_command()
     async def next(self, ctx):
-        if len(self.player_queue) > 0:
-            player = self.player_queue[0]
-            source = self.source_queue[0]
-            ctx.voice_client.stop()
-            await ctx.respond(f'Now playing: ***{player.title}***\nFrom {source}')
-        else:
-            await ctx.respond('Queue is empty')
+        voice_client = self.voice_clients[ctx.guild.id]
+        change = False
+        if voice_client.loop:
+            voice_client.toggle_loop()
+            change = True
+        voice_client.voice.stop()
+        if change:
+            voice_client.toggle_loop()
 
     @commands.slash_command()
     async def queue(self, ctx):
-        queue_len = len(self.player_queue)
-        if queue_len == 0:
+        voice_client = self.voice_clients[ctx.guild.id]
+        if voice_client.is_empty():
             await ctx.respond('Queue is empty')
         else:
             embed = discord.Embed(title='Queue')
-            for i in range(queue_len):
-                embed.add_field(name=str(i), value=self.player_queue[i].title, inline=False)
+            for i in range(len(voice_client.player_queue)):
+                embed.add_field(name=f'{i} (currently playing)' if i == 0 else str(i), value=voice_client.player_queue[i].title, inline=False)
             await ctx.respond(embed=embed)
-
-    @commands.slash_command()
-    async def purge(self, ctx):
-        self.player_queue, self.source_queue = ([], [])
-        if ctx is not None:
-            await ctx.respond('Purged queue')
 
     @commands.slash_command()
     async def download(self, ctx, *, url):
@@ -228,10 +299,11 @@ class Voice(commands.Cog):
 
     @fart.before_invoke
     async def ensure_voice(self, ctx):
+        # TODO refaire ça, ça va pas du tout
         if ctx.voice_client is None:
             await self.join(ctx, None, False)
         try:
-            ctx.voice_client.source.volume = self.volume
+            ctx.voice_client.source.volume = self.default_volume
         finally:
             return
